Add a path to warn message when type is invalid
When you have thousands of files to compress and it fails with `srcFile should be a valid file or directory`, go try to figure which one and why. This changeset gives a relief to end user by at least pointing to file (path) as a start point for further investigation.